requirements.txt
Pillow>=9.0      # for Image, ImageTk
requests>=2.0    # for the optional WTF image lookup